# packages/api-contracts/openapi/wishpool.yaml
openapi: 3.1.0
info:
  title: WishPool API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    WishPool core API contract for mobile apps, parent web, admin web, and backend services.
    All write endpoints are command-oriented and should be called with an Idempotency-Key header.
servers:
  - url: http://localhost:8080
    description: Local core-api
security:
  - bearerAuth: []
...
paths:
...
  /internal/workflows/materialize-weekly-plan:
    post:
      tags: [Internal]
      security:
        - internalToken: []
      operationId: runMaterializeWeeklyPlanWorkflowActivity
      summary: Execute weekly plan materialization activity.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterializeWeeklyPlanWorkflowRequest'
      responses:
        '200':
          description: Weekly plan materialized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyPlan'
        default:
          $ref: '#/components/responses/Error'
  /internal/workflows/evaluate-reward:
    post:
      tags: [Internal]
      security:
        - internalToken: []
      operationId: runRewardEvaluationWorkflowActivity
      summary: Execute reward evaluation activity.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluateRewardWorkflowRequest'
      responses:
        '200':
          description: Reward evaluation completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowAcceptedResponse'
        default:
          $ref: '#/components/responses/Error'
  /internal/workflows/run-ai-precheck:
    post:
      tags: [Internal]
      security:
        - internalToken: []
      operationId: runAiPrecheckWorkflowActivity
      summary: Execute submission AI precheck activity.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunAiPrecheckWorkflowRequest'
      responses:
        '200':
          description: AI precheck completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowAcceptedResponse'
        default:
          $ref: '#/components/responses/Error'
  /internal/workflows/generate-memory:
    post:
      tags: [Internal]
      security:
        - internalToken: []
      operationId: acceptGenerateMemoryWorkflowActivity
      summary: Execute weekly memory workflow activity.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateMemoryWorkflowRequest'
      responses:
        '200':
          description: Memory workflow activity completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowAcceptedResponse'
        default:
          $ref: '#/components/responses/Error'
  /internal/workflows/privacy-deletion:
    post:
      tags: [Internal]
      security:
        - internalToken: []
      operationId: acceptPrivacyDeletionWorkflowActivity
      summary: Execute privacy deletion workflow activity.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrivacyDeletionWorkflowRequest'
      responses:
        '200':
          description: Privacy deletion workflow activity completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowAcceptedResponse'
        default:
          $ref: '#/components/responses/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
...
  responses:
    Error:
      description: Error response.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      required: [type, title, status, traceId]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        traceId:
          type: string
...
    TaskCategory:
      type: string
      enum: [study, reading, exercise, habit, custom]
    SubmissionType:
      type: string
      enum: [photo, audio, video, manual]
...
    WeeklyPlanRuleInput:
      type: object
      required: [title, category, submissionType, weekdays, isCore, requireReview]
      properties:
        taskTemplateId:
          anyOf:
            - type: string
              format: uuid
            - type: 'null'
        title:
          type: string
        category:
          $ref: '#/components/schemas/TaskCategory'
        submissionType:
          $ref: '#/components/schemas/SubmissionType'
        description:
          type: string
        targetText:
          type: string
        weekdays:
          type: array
          items:
            type: integer
            minimum: 1
            maximum: 7
        isCore:
          type: boolean
        requireReview:
          type: boolean
        sortOrder:
          type: integer
    WeeklyPlan:
      type: object
      required: [id, familyId, childId, weekId, startDate, endDate, rewardMode, status, rules]
      properties:
        id:
          type: string
          format: uuid
        familyId:
          type: string
          format: uuid
        childId:
          type: string
          format: uuid
        weekId:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        rewardMode:
          type: string
          enum: [flexible, strict]
        status:
          type: string
          enum: [draft, active, archived]
        rules:
          type: array
          items:
            $ref: '#/components/schemas/WeeklyPlanRule'
    WeeklyPlanRule:
      allOf:
        - $ref: '#/components/schemas/WeeklyPlanRuleInput'
        - type: object
          required: [id]
          properties:
            id:
              type: string
              format: uuid
...
    MaterializeWeeklyPlanWorkflowRequest:
      type: object
      required: [weeklyPlanId]
      properties:
        weeklyPlanId:
          type: string
          format: uuid
        triggeredByEventId:
          type: string
          format: uuid
    EvaluateRewardWorkflowRequest:
      type: object
      required: [eventType, taskInstanceId, actorUserId]
      properties:
        eventType:
          type: string
          enum: [review.approved, task.skipped, review.revoked]
        taskInstanceId:
          type: string
          format: uuid
        reviewId:
          type: string
          format: uuid
        actorUserId:
          type: string
          format: uuid
        triggeredByEventId:
          type: string
          format: uuid
    RunAiPrecheckWorkflowRequest:
      type: object
      required: [submissionId]
      properties:
        submissionId:
          type: string
          format: uuid
        triggeredByEventId:
          type: string
          format: uuid
    GenerateMemoryWorkflowRequest:
      type: object
      required: [triggeredByEventId]
      properties:
        triggeredByEventId:
          type: string
          format: uuid
    PrivacyDeletionWorkflowRequest:
      type: object
      required: [triggeredByEventId]
      properties:
        triggeredByEventId:
          type: string
          format: uuid
    WorkflowAcceptedResponse:
      type: object
      required: [workflow, status, acceptedAt]
      properties:
        workflow:
          type: string
        status:
          type: string
        acceptedAt:
          type: string
          format: date-time
...